feat: add validator dependency
