+++ package.json
@@ -56,6 +56,7 @@
     "redux-thunk": "2.4.1",
     "ttag": "1.8.6",
     "unleash-proxy-client": "3.2.0",
+    "validator": "^13.12.0",
     "viz.js": "2.1.2"
   },
   "main": "public/electron.js",
