+++ src/reducers/index.js
@@ -15,7 +15,6 @@ import { PROPOSAL_DOWNLOAD_STATUS } from '../utils/atomicSwap';
 import { HATHOR_TOKEN_CONFIG } from "@hathor/wallet-lib/lib/constants";
 import helpersUtils from '../utils/helpers';
 import LOCAL_STORE from '../storage';
-import { getGlobalWallet } from "../modules/wallet";
 
 /**
  * @typedef TokenHistory
@@ -258,8 +257,8 @@ const rootReducer = (state = initialState, action) => {
       return onUpdateTxHistory(state, action);
     case types.WALLET_CHANGE_STATE:
       return onWalletStateChanged(state, action);
-    case types.UPDATE_MINING_SERVER:
-      return onUpdateMiningServer(state, action);
+    case types.SET_MINING_SERVER:
+      return onSetMiningServer(state, action);
     default:
       return state;
   }
@@ -1008,23 +1007,9 @@ export const onWalletStateChanged = (state, { payload }) => ({
   walletState: payload,
 });
 
-export const onUpdateMiningServer = (state, { payload }) => {
-  const wallet = getGlobalWallet();
-  if (payload.reset) {
-    wallet.storage.config.setTxMiningUrl();
-    LOCAL_STORE.resetMiningServer();
-    return {
-      ...state,
-      miningServer: null,
-    }
-  } else {
-    wallet.storage.config.setTxMiningUrl(payload.url);
-    LOCAL_STORE.setMiningServer(payload.url);
-    return {
-      ...state,
-      miningServer: payload.url,
-    }
-  }
-}
+export const onSetMiningServer = (state, { payload }) => ({
+  ...state,
+  miningServer: payload,
+});
 
 export default rootReducer;
